% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nldjhalf.R
\name{add}
\alias{add}
\title{Toplama Fonksiyonu}
\usage{
add(x, y)
}
\arguments{
\item{x}{Birinci sayı.}

\item{y}{İkinci sayı.}
}
\value{
İki sayının toplamı.
}
\description{
Bu fonksiyon iki sayıyı toplar.
}
