% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nldjhalf.R
\name{multiply}
\alias{multiply}
\title{Çarpma Fonksiyonu}
\usage{
multiply(x, y)
}
\arguments{
\item{x}{Birinci çarpan.}

\item{y}{İkinci çarpan.}
}
\value{
İki çarpanın çarpımı.
}
\description{
Bu fonksiyon iki sayıyı çarpar.
}
